requirements.txt
aiohttp==3.5.4
coloredlogs==10.0
Cython==0.29.13
Django==2.2.4
django-allauth==0.40.0
django-extensions==2.2.1
jsonpickle==1.2
mysqlclient==1.4.4
requests==2.22.0
SQLAlchemy==1.3.7
websockets==8.0.2
